Improve compatibility with BTT Display
Board and Schematic changed to accept a reset signal from either the "Reset" line or the "Kill" line and changes the default configuration to use the "Reset" line.  Big Tree Tech boards are configured by default to use the reset line and Fysetc boards are configured by default to use both the reset and kill lines.  The proposed changes should allow the reset button on either the Fysetc or the BTT display to work without requiring the user to reconfigure the adapter or display.  It also gives the option to reconfigure the adapter to use the "Kill" line or to disable the function if desired.